requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-genai>=0.5.0
google-adk>=0.1.3
google-api-python-client>=2.137.0
google-auth>=2.32.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
crawl4ai>=0.3.9
uvicorn
playwright